requirements.txt
gevent
certifi
chardet
click
Flask
gunicorn
idna
itsdangerous
Jinja2
MarkupSafe
PyYAML
requests
urllib3
Werkzeug
six
